# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "calvin3oo-cse-341.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/contacts/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "body format",
            "schema": {
              "$ref": "#/definitions/addUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/{docID}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "docID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "docID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "body format",
            "schema": {
              "$ref": "#/definitions/updateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "docID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "addUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "name"
        },
        "lastName": {
          "type": "string",
          "example": "last"
        },
        "email": {
          "type": "string",
          "example": "email"
        },
        "favoriteColor": {
          "type": "string",
          "example": "color"
        },
        "birthday": {
          "type": "string",
          "example": "birthday"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ]
    },
    "updateUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "name"
        },
        "lastName": {
          "type": "string",
          "example": "last"
        },
        "email": {
          "type": "string",
          "example": "email"
        },
        "favoriteColor": {
          "type": "string",
          "example": "color"
        },
        "birthday": {
          "type": "string",
          "example": "birthday"
        }
      }
    }
  }
}